+++ Core/Misc.cs
@@ -451,6 +451,10 @@ namespace PortAIO.Dual_Port
 
             dutility.AddItem(new MenuItem("enablePredictioner", "Enable Predictioner?").SetValue(false));
             dutility.AddItem(new MenuItem("Predictioner", "Which Predictioner?").SetValue(new StringList(new[] { "SPredictioner", "OKTWPredictioner" })));
+
+            dutility.AddItem(new MenuItem("enableCamera", "Enable Camera-Mod?").SetValue(false));
+            dutility.AddItem(new MenuItem("Camera", "Which Camera-Mod?").SetValue(new StringList(new[] { "HSCamera" })));
+
             menu.AddSubMenu(dutility);
 
             var utility = new Menu("Standalone Utilities", "PortAIOuTILITIESS");
